{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "{{cookiecutter.version}}"
description = "{{cookiecutter.description}}"
authors = ["{{cookiecutter.author_name}}"]

[[tool.poetry.source]]
name = "wcc-pypi"
url = "https://pypi.wochacha.cn/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.11"

pytz = "^2022.7" # https://github.com/stub42/pytz
python-slugify = "^7.0.0" # https://github.com/un33k/python-slugify
Pillow = "^9.4.0" # https://github.com/python-pillow/Pillow
argon2-cffi = "^21.3.0" # https://github.com/hynek/argon2_cffi
redis = "^4.4.1" # https://github.com/redis/redis-py
hiredis = "^2.1.0" # https://github.com/redis/hiredis-py
celery = { extras = [
  "amqp",
  "gevent",
], version = "^5.2.7" } # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat = ">=2.4.0" # https://github.com/celery/django-celery-beat
httpx = { extras = [
  "http2",
], version = "^0.22.0" } # https://github.com/encode/httpx/

# Django
# ------------------------------------------------------------------------------
django = "^4.1.5" # pyup: < 5.0  # https://www.djangoproject.com/
django-environ = "^0.9.0" # https://github.com/joke2k/django-environ
django-model-utils = "^4.3.1" # https://github.com/jazzband/django-model-utils
django-crispy-forms = "^1.14.0" # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5 = "^0.7" # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis = "^5.2.0" # https://github.com/jazzband/django-redis
channels = "^3.0.4" # https://github.com/django/channels
channels-redis = "^3.3.1" # https://github.com/django/channels_redis
django-storages = { version = "^1.13.2", extras = [
  "boto3",
] } # https://github.com/jschneier/django-storages
django-admin-interface = ">=0.19.1" # https://github.com/fabiocaccamo/django-admin-interface
wcc-python-sdk-django-casdoor = "^1.0.6" # https://github.com/wccdev/django-casdoor-auth

# Django REST Framework
djangorestframework = "^3.14.0"       # https://github.com/encode/django-rest-framework
django-cors-headers = "^3.13.0"       # https://github.com/adamchainz/django-cors-headers
django-user-notification = ">=0.7.11" # https://github.com/anyidea/django-user-notification
drfexts = "^0.19.6"                   # https://github.com/anyidea/drfexts
drf-flex-fields = ">=0.9.7"           # https://github.com/rsinger86/drf-flex-fields
drf-spectacular = ">=0.25.1"          # https://github.com/tfranzel/drf-spectacular

[tool.poetry.group.production]
optional = true

[tool.poetry.group.production.dependencies]
uvicorn = { version = "^0.18.2", extras = [
  "standard",
] } # https://github.com/encode/uvicorn
psycopg2 = "^2.9.5" # https://github.com/psycopg/psycopg2
sentry-sdk = "^1.12.1" # https://github.com/getsentry/sentry-python
gunicorn = "^20.1.0" # https://github.com/benoitc/gunicorn

# Django
# ------------------------------------------------------------------------------
django-anymail = "^9.0" # https://github.com/anymail/django-anymail
collectfast = "^2.2.0"  # https://github.com/antonagestam/collectfast

# Django REST Framework
# ------------------------------------------------------------------------------

[tool.poetry.group.dev.dependencies]
Werkzeug = ">=2.2.2"       # https://github.com/pallets/werkzeug
ipdb = ">=0.13.11"         # https://github.com/gotcha/ipdb
bpython = "^0.23"          # https://github.com/bpython/bpython
psycopg2-binary = "^2.9.5" # https://github.com/psycopg/psycopg2
watchfiles = ">=0.18.1"    # https://github.com/samuelcolvin/watchfiles

# Testing
# ------------------------------------------------------------------------------
mypy = ">=0.982"                      # https://github.com/python/mypy
django-stubs = ">=1.13.1"             # https://github.com/typeddjango/django-stubs
pytest = ">=7.2.0"                    # https://github.com/pytest-dev/pytest
pytest-sugar = ">=0.9.6"              # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs = ">=1.8.0" # https://github.com/typeddjango/djangorestframework-stubs

# Documentation
# ------------------------------------------------------------------------------
sphinx = ">=5.3.0"               # https://github.com/sphinx-doc/sphinx
sphinx-autobuild = ">=2021.3.14" # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8 = ">=6.0.0"        # https://github.com/PyCQA/flake8
flake8-isort = ">=6.0.0"  # https://github.com/gforcada/flake8-isort
toml = ">=0.10.2"         # https://github.com/uiri/toml
coverage = ">=7.0.4"      # https://github.com/nedbat/coveragepy
black = ">=22.12.0"       # https://github.com/psf/black
pylint-django = ">=2.5.3" # https://github.com/PyCQA/pylint-django
pylint-celery = ">=0.3"   # https://github.com/PyCQA/pylint-celery
pre-commit = ">=2.21.0"   # https://github.com/pre-commit/pre-commit
commitizen = ">=2.28.0"   # https://github.com/commitizen-tools/commitizen

# Django
# ------------------------------------------------------------------------------
factory-boy = ">=3.2.1"            # https://github.com/FactoryBoy/factory_boy
django-debug-toolbar = ">=3.8.1"   # https://github.com/jazzband/django-debug-toolbar
django-extensions = ">=3.2.1"      # https://github.com/django-extensions/django-extensions
django-coverage-plugin = ">=3.0.0" # https://github.com/nedbat/django_coverage_plugin
pytest-django = ">=4.5.2"          # https://github.com/pytest-dev/pytest-django

# Django REST Framework
# ------------------------------------------------------------------------------

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations

  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
